requirements.txt
streamlit
transformers
Pillow
torch

logging
